Switch to lowercase keyspace name.
Must be my version of cassandra (1.1.5), but I could not get the tests to run without changing to a lowercase keyspace name.

I tried just doing the include check with lowercase, but cassandra really did not like the switching of case in various places. This seems to work best from what I can tell.